add decrepit namespace and put SEED, IDEA, Blowfish, and CAST5 in it
