+++ tests/hazmat/primitives/decrepit/__init__.py
@@ -1,0 +1,3 @@
+# This file is dual licensed under the terms of the Apache License, Version
+# 2.0, and the BSD License. See the LICENSE file in the root of this repository
+# for complete details.
